+++ examples/ng-example/ios/.gitignore
@@ -1,0 +1,13 @@
+# NPM renames .gitignore to .npmignore
+# In order to prevent that, we remove the initial "."
+# And the CLI then renames it
+
+App/build
+App/Pods
+App/public
+App/Podfile.lock
+xcuserdata
+
+# Cordova plugins for Capacitor
+capacitor-cordova-ios-plugins
+
