+++ examples/ng-example/ios/App/App/capacitor.config.json
