+++ examples/ng-example/src/app/home/home.page.html
@@ -9,12 +9,16 @@
 <ion-content [fullscreen]="true">
   <ion-header collapse="condense">
     <ion-toolbar>
-      <ion-title size="large">Blank</ion-title>
+      <ion-title size="large">ng-example</ion-title>
     </ion-toolbar>
   </ion-header>
 
   <div id="container">
-    <strong>Ready to create an app?</strong>
-    <p>Start with Ionic <a target="_blank" rel="noopener noreferrer" href="https://ionicframework.com/docs/components">UI Components</a></p>
+    <h2>Status: <span>{{authSuccess ? 'Authenticated' : 'Unauthenticated'}}</span></h2>
+    <pre *ngIf="userProfile">{{userProfile | json}}</pre>
+
+    <div *ngIf="authSuccess"></div>
+    <ion-button (click)="login()">Login</ion-button>
+    <ion-button (click)="logout()">Logout</ion-button>
   </div>
 </ion-content>
